Added Uniqueness of board checking function
Each finished puzzle needs to be checked for uniqueness, or else even
though the user submits the right board, it is considered wrong. Here I
just ran the puzzle maker till the board was unique and every 10 tries I
changed the filled in board itself. The minimum number of values still
in the board has to be 28 or else the script will time out. Will need to
implement a queue either in program memory or maintained on the database
I have to manage rn, will take 3 and a half hours (roughly) to compute
so it is probably gonna be stored in the DB. I have to eat lunch.